Initial plugin logic
This plugin exposes some more of the logic we were sharing via
https://github.com/grapl-security/buildkite-common as a formal
Buildkite plugin.

In particular, this commit introduces the logic for updating our
internal tracking tags, as well as the diff script we use (which
leverages those same tags).

For now, this is a "library" plugin, which basically allows us to
inject scripts and binaries into the `PATH`, which commands can then
use as they see fit.

Longer-term, we'll add more release-centric logic to this plugin.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>