Add "real run" awareness to tag push operation
This adds the `is_real_run` logic we use in our `grapl-rc` Buildkite
plugin to this plugin. If we're doing a "test run" of any pipeline
that uses this plugin, we don't want to inadvertently push tags to
Github; that should only happen for "real" pipeline runs (here, "real"
is currently defined as any build not triggered via the UI).

Signed-off-by: Christopher Maier <chris@graplsecurity.com>